관리자 권한을 가진 계정은 학습 자료를 수정할 수 있는 기능 추가 (#95)
보상금액 자료형을 Double로 변경함으로써 소수점 데이터를 처리할 수 있게되었음. 관리자는 미션과 아티클을 수정할 수 있는 기능 추가함

+++ back-end/src/main/java/world/ludium/education/profile/dto/MyMissionDTO.java
@@ -10,7 +10,7 @@ public record MyMissionDTO(UUID missionId,
                            UUID postingId,
                            UUID curriculumId,
                            UUID rewardToken,
-                           Integer rewardAmount,
+                           Double rewardAmount,
                            String rewardClaimStatus) {
 
 }
